% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{fixedb.test}
\alias{fixedb.test}
\title{Fixed-b unit root test}
\usage{
fixedb.test(y, B = floor(0.2 * length(y)), HC = TRUE)
}
\arguments{
\item{y}{Vector to be tested for a unit root.}

\item{B}{An integer defining the blocklength; default is \eqn{0.2*T}.}

\item{HC}{A logical variable defining whether the heteroskedasticity robust version (\eqn{HC = TRUE})
or the homoskedasticity version (\eqn{HC = FALSE}) should be used; default is \eqn{HC = TRUE}..}
}
\value{
A list containing the following components:
\item{teststatistic}{The value of the small-b test statistic.}
\item{crit.values}{A vector of critical valiues for different significance levels.}
\item{rejection}{A logical vector indicating whether the unit root hypothesis is rejected in favor of stationarity for at different significance levels.}
\item{crit.blocklength}{The relative blocklength for which the critical values are reported.}
}
\description{
Performs the fixed-b unit root test
}
\examples{
series1 <- rnorm(100) + c(rep(0,60), rep(5, 40))
fixedb.test(series1)
series2 <- cumsum(rnorm(100)) + c(rep(0,60), rep(5, 40))
fixedb.test(series2)
}
