% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{get.prewhitened}
\alias{get.prewhitened}
\title{Prewhitened series}
\usage{
get.prewhitened(y, p = 1)
}
\arguments{
\item{y}{Vector of time series values to be pre-whitened.}

\item{p}{Lag length of the AR model.}
}
\value{
A vector containing the values of the pre-whitened series.
}
\description{
Returns the prewhitened series given some lag length.
}
\examples{
get.prewhitened(rnorm(100))
get.prewhitened(arima.sim(list(order=c(1,0,1), ar=.5, ma=.0), n = 100))
}
