% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{lagselection.BIC}
\alias{lagselection.BIC}
\title{BIC laglegth selection}
\usage{
lagselection.BIC(y, Pmax = 5)
}
\arguments{
\item{y}{Vector of time series values to be pre-whitened.}

\item{Pmax}{Maximum lag length to be considered.}
}
\value{
Lag length determined by the BIC criterion.
}
\description{
Returns the laglength in the prewhitening regression for which the BIC criterion is minimized.
}
\examples{
lagselection.BIC(rnorm(100))
}
