% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{smallb.test}
\alias{smallb.test}
\title{Small-b unit root test}
\usage{
smallb.test(y, B = floor(length(y)^0.7), HC = TRUE)
}
\arguments{
\item{y}{Vector to be tested for a unit root.}

\item{B}{An integer defining the blocklength; default is \eqn{T^{0.7}}.}

\item{HC}{A logical variable defining whether the heteroskedasticity robust version (\eqn{HC = TRUE}) 
or the homoskedasticity version (\eqn{HC = FALSE}) should be used; default is \eqn{HC = TRUE}.}
}
\value{
A list containing the following components:
\item{teststatistic}{The value of the small-b test statistic.}
\item{p.value}{The p-value of the small-b test against stationarity.}
}
\description{
Performs the small-b unit root test
}
\examples{
series1 <- rnorm(100) + c(rep(0,60), rep(5, 40))
smallb.test(series1)
series2 <- cumsum(rnorm(100)) + c(rep(0,60), rep(5, 40))
smallb.test(series2)
## autocorrelation robust test
series3 <- filter(rnorm(300),0.8,"recursive")
series3.pw <- get.prewhitened(series3, p = 1)
smallb.test(series3.pw)
}
